+++ packages/style-switcher/package.json
@@ -14,8 +14,8 @@
 		"build": "svelte-kit sync && svelte-package",
 		"check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
 		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
-		"lint": "prettier --plugin-search-dir . --check . && eslint .",
-		"format": "prettier --plugin-search-dir . --write ."
+		"lint": "prettier --check . && eslint .",
+		"format": "prettier --write ."
 	},
 	"repository": {
 		"type": "git",
@@ -35,17 +35,19 @@
 	"homepage": "https://github.com/watergis/svelte-maplibre-components/tree/main/packages/style-switcher#readme",
 	"devDependencies": {
 		"@creativebulma/bulma-tooltip": "^1.2.0",
-		"@sveltejs/adapter-auto": "^2.1.1",
-		"@sveltejs/kit": "^1.30.4",
+		"@sveltejs/adapter-auto": "^3.1.1",
+		"@sveltejs/kit": "^2.5.4",
 		"@sveltejs/package": "^2.3.0",
+		"@sveltejs/vite-plugin-svelte": "^3.0.2",
 		"@types/debounce": "^1.2.4",
 		"@types/lodash-es": "^4.17.12",
-		"@typescript-eslint/eslint-plugin": "^6.21.0",
-		"@typescript-eslint/parser": "^6.21.0",
+		"@typescript-eslint/eslint-plugin": "^7.2.0",
+		"@typescript-eslint/parser": "^7.2.0",
 		"@watergis/svelte-maplibre-menu": "workspace:^",
 		"eslint": "^8.57.0",
 		"eslint-config-prettier": "^9.1.0",
 		"eslint-plugin-svelte": "^2.35.1",
+		"pmtiles": "^3.0.5",
 		"prettier": "^3.2.5",
 		"prettier-plugin-svelte": "^3.2.2",
 		"sass": "^1.72.0",
@@ -54,12 +56,12 @@
 		"svelte-preprocess": "^5.1.3",
 		"tslib": "^2.6.2",
 		"typescript": "^5.4.2",
-		"vite": "^4.5.2"
+		"vite": "^5.1.6"
 	},
 	"type": "module",
 	"dependencies": {
-		"debounce": "^1.2.1",
-		"maplibre-gl": "^3.6.2"
+		"debounce": "^2.0.0",
+		"maplibre-gl": "^4.1.0"
 	},
 	"peerDependencies": {
 		"svelte": "^4.0.0"
