--- packages/style-switcher/src/routes/+page.ts
@@ -1,2 +1,0 @@
-export const prerender = true;
-export const ssr = false;
